+++ POLEGON/POLEGON/random_utils.hpp
@@ -18,6 +18,6 @@
 extern std::mt19937 random_engine;
 extern std::uniform_real_distribution<> uniform_distribution;
 
-float uniform_random();
+double uniform_random();
 
 #endif /* random_utils_hpp */
